+++ test-data/cljfmt-sabndbox.clj
@@ -6,6 +6,10 @@
   (let [a b
         aa bb
         ccc {:a b :aa bb :ccc ccc}]))
+(tabular
+ (foo
+  (bar
+   0)))
 
 (f
  ##Inf)